Landmark commit: Straightline code for datalink layer should be complete, minus timers.
Nonfuntional timer code was stripped from this commit. Timers have proven to be somewhat difficult to integrate.

Small error of note: Data frames are always sent as 150 bytes, never smaller. Trivial to fix, but not included.
(Fix is untested).

This code is currently written for Go-Back-1. However, it should work perfectly for go-back-N with a trivial change.
Just set the #define in datalink_layer.h to 4. This is completely untested and may break everything.

Error introduction code *has not yet been introduced* at this juncture. So all frames get sent perfectly.